Split annotation package into api and impl
