+++ src/main/java/me/zeroeightsix/fiber/api/annotation/Listener.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.annotation;
+package me.zeroeightsix.fiber.api.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
