+++ src/test/java/me/zeroeightsix/fiber/annotation/AnnotatedSettingsTest.java
@@ -1,5 +1,9 @@
 package me.zeroeightsix.fiber.annotation;
 
+import me.zeroeightsix.fiber.api.annotation.AnnotatedSettings;
+import me.zeroeightsix.fiber.api.annotation.Listener;
+import me.zeroeightsix.fiber.api.annotation.Setting;
+import me.zeroeightsix.fiber.api.annotation.Settings;
 import me.zeroeightsix.fiber.builder.ConfigTreeBuilder;
 import me.zeroeightsix.fiber.exception.FiberException;
 import me.zeroeightsix.fiber.tree.ConfigTree;
